+++ app/src/main/java/ajitsingh/com/expensemanager/activity/CurrentWeekExpenseFragment.java
@@ -43,7 +43,7 @@ public class CurrentWeekExpenseFragment extends Fragment implements CurrentWeekE
   }
 
   @Override
-  public void displayTotalExpenses(Long totalExpense) {
+  public void displayTotalExpenses(Float totalExpense) {
     TextView totalExpenseTextBox = (TextView) getActivity().findViewById(R.id.current_week_expense);
     totalExpenseTextBox.setText(getString(R.string.total_expense) + " " + getString(R.string.rupee_sym) + totalExpense);
   }
