+++ app/src/main/java/ajitsingh/com/expensemanager/view/TodaysExpenseView.java
@@ -5,6 +5,6 @@ import java.util.List;
 import ajitsingh.com.expensemanager.model.Expense;
 
 public interface TodaysExpenseView {
-  void displayTotalExpense(Long totalExpense);
+  void displayTotalExpense(Float totalExpense);
   void displayTodaysExpenses(List<Expense> expenses);
 }
